refactor: permission model
