+++ interfaces/index.ts
@@ -2,3 +2,4 @@ export * from "./category"
 export * from "./supplier"
 export * from "./user"
 export * from "./rolePermission"
+export * from "./permission"
